refactor: StorageService를 interface로 변경
